feat: Created a new internal controller for refresh role assignment
